backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pydantic>=1.8.2
python-dotenv>=0.19.0
langchain>=0.0.267
langchain-groq>=0.3.0
